requirements.txt
SQLAlchemy
python-jose
pydantic[email]
pytz
psycopg2-binary
requests
python-telegram-bot
dotenv